Inclusão de testes

+++ src/main/java/br/com/fiap/PetPass/controller/EstabelecimentoController.java
@@ -3,6 +3,8 @@ package br.com.fiap.PetPass.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.com.fiap.PetPass.dto.EstabelecimentoDTO;
+import br.com.fiap.PetPass.entity.Estabelecimento;
 import br.com.fiap.PetPass.service.EstabelecimentoService;
 
 @RestController
@@ -68,8 +71,9 @@ public class EstabelecimentoController {
 	}
 	
 	@PostMapping
-	public EstabelecimentoDTO create(@RequestBody EstabelecimentoDTO estabelecimento) {
-		return this.service.create(estabelecimento);
+	public ResponseEntity<EstabelecimentoDTO> create(@RequestBody EstabelecimentoDTO estabelecimento) {
+		EstabelecimentoDTO estabelecimentoDTO = this.service.create(estabelecimento);
+		return ResponseEntity.status(HttpStatus.CREATED).body(estabelecimentoDTO);
 	}
 	
 	@PutMapping
@@ -77,9 +81,16 @@ public class EstabelecimentoController {
 		return this.service.update(estabelecimento);
 	}
 	
-	@DeleteMapping
-	public void delete(@PathVariable Long id) {
+	@DeleteMapping("{id}")
+	public ResponseEntity<?> delete(@PathVariable Long id) {
+		EstabelecimentoDTO estabelecimento = service.findById(id);
+		
+		if(estabelecimento == null) {
+			return ResponseEntity.notFound().build();
+		}
+	
 		this.service.delete(id);
+		return ResponseEntity.ok(id);
 	}
 	
 }
